+++ codegen/tests/fail/delegate/external_trait_wrong_declaration.rs
@@ -1,0 +1,34 @@
+use delegation::delegate;
+
+#[delegate(as = "AsRef")]
+trait AsRefDef<T: ?Sized> {
+    fn as_ref(&mut self) -> &mut T;
+}
+
+#[delegate]
+trait AsStr {
+    fn as_str(&self) -> &str;
+}
+
+impl AsStr for String {
+    fn as_str(&self) -> &str {
+        self
+    }
+}
+
+#[delegate(as = "AsStr")]
+trait AsStrDef {
+    fn as_str(&self) -> &str;
+}
+
+#[delegate(derive(
+for<> AsRef<str> as AsRefDef;
+AsStr as AsStrDef;
+))]
+enum Name {
+    First(String),
+}
+
+fn main() {
+    unreachable!()
+}