+++ codegen/tests/fail/delegate/newtype_inner_not_impls_trait.rs
@@ -1,0 +1,13 @@
+use delegation::delegate;
+
+#[delegate]
+trait AsStr {
+    fn as_str(&self) -> &str;
+}
+
+#[delegate(derive(AsStr))]
+struct FirstName(String);
+
+fn main() {
+    unreachable!()
+}