+++ codegen/tests/fail/delegate/not_allowed_macro_in_trait_definition.rs
@@ -1,0 +1,10 @@
+use delegation::delegate;
+
+#[delegate]
+trait Channel {
+    unreachable!();
+}
+
+fn main() {
+    unreachable!()
+}