+++ codegen/tests/fail/delegate/struct_with_multiple_fields.rs
@@ -1,0 +1,14 @@
+use delegation::delegate;
+
+#[delegate]
+struct FirstName(String, String);
+
+#[delegate]
+struct Name {
+    first: String,
+    last: String,
+}
+
+fn main() {
+    unreachable!()
+}