+++ CHANGELOG.md
@@ -18,6 +18,17 @@ All user visible changes to this project will be documented in this file. This p
     - `External` and `Wrapper` structs (#1);
     - `Convert` and `TypeOf` traits (#1).
 
+### Added
+
+- `#[delegate]` macro: ([#1])
+    - Single-fielded structs support.
+    - Enums with single-fielded variants support.
+    - Limited generics support.
+    - External types support.
+    - External traits support.
+
+[#1]: https://github.com/arcane-rs/delegation/pull/1
+